Composing New Component for Section in Earn Page
